Add rules to the cross terminal enum.
We make sure that we separate the non terminals from the terminals.
The general program structure is in place, but there are some non-terminals
still missing. I'll look into it later. So far everything compiles and the
tests work fine.

The changes to parser/dependency/CrossTerminalParser.hpp merely test to
see if adding crossterminals works. Later one, the grammar will be
defined in this class's constructor.