Fix a memory leak in the parser.
This trivial leak is caused by the parser not deallocating
the pointer to the syntax tree root upon deallocation.
For getting the syntax tree, we should probably null the root
and forget about it... Maybe it's better to make the entire
root an std::unique_ptr. This will be considered for future
updates.