Implement argument lists in functions
This implies argument lists that are
continuations from previous types. As
an example:

`(32u a_, b_, c_ : )`

is a perfectly valid signature. However,
also
`(32u a_, b_, 16s c_ : )`

is completely valid. The semantics are
that a_ and b_ are of type 32u, whilst
c_ is of type 16s. This works for any
type.