Add var keyword for data declaration statements inside functions.
The statement ambiguity in LL(1) forces us to split up our parsing inside
a partial expression of the form

	a.Aa

What comes after the last 'a'? If it is a dot, them we have an expression.
If it is a space, we have a declaration. LL(1) can not ascertain this
cleanly and will create an extremely ugly and flawed parse tree for this
specific entity.

It is also more easily spotted what is a non-expression statement. This can
be very useful for readability. I don't see a problem with the "var" keyword.

The new usage is

	var a.Aa

this lets the parser - as well as the programmer - know that we wish to declare
an object/instance/POD of the following type. This must also be used in front of
_any_ type, even if 'const' or 'ref' are disambiguous. The reason for this is to
avoid special cases.