Remove the groupers and symbols from the abstract syntax tree generator.
It's worth noting that it may be beneficial to perform this step already in
the parser itself. When it notices that a token's type (token_type) equates
GROUPER_SYMBOL or OTHER_SYMBOL, that the deduction rule does not put those
types into the tree. They must be pushed to the stack, but to put them into
the tree is both unnecessary and wasteful. (Don't prematurely optimize).