Add a block-less syntax for if, else, while,.. and solve the dangling else problem.
The dangling else problem is solved by pairing it with the nearest if. A detached
else would be nice to have. I also have to add static ifs! I wonder how a the
detached else would work... it could be a statement, which would ambiguate which
if to prefer.

Let's think a little...
Suppose you have the following code.

	if (a_)
		doSomething();
	a_ += 1;
	else
		doSomethingElse();

It would be cool if the else would correctly fire here. How would the parse-tree
look like for this specific sequence?

Statement list
	if statement
		a_
		doSomething
	Statement list
		a += 1;
		Statement list
			else statement
				doSomethingElse();

Flattening this gives us

Statement list
	if statement
		a_
		doSomething()
	expression
		a
		+=
		1
	else statement
		doSomethingElse();

The semantic analyzer needs to keep track of the last if. When a new non-else
statement comes directly after, it sets a value/ptr that points to that if
statement. If a stray else is found, we can go back to the if and inject a boolean
there.