Massive refactoring according to the ULCRI scheme.
The dependencies are now as stated in the rationale.
There will be some smaller changes, such as docs and
the dependency graph.

This commit localized LLOneProductionDeducer and the
Mealy machine where they are actually used. This
decreases the apparent global dependencies. It's
possible (probably) to do the same for TokenTypeToCrossTerminal.
That one's only used in the lexer, I think..

Ok so it's used by the TreeBuilder as well as the Parser.
The TreeBuilder's usage of it doesn't appear anywhere.
I'll be deleting it and then making it part of the parser
alone. That should keep the global dependencies to an
absolute minimum.