Move the semantic analyzer into its own folder.
The project has to keep a state that is going to be used for the language itself.
Right now, testing the tree pruner is not really useful without parser output.
The unit test is supposed to be singular, and have no more dependencies (except
IO compared to the thing that it is testing.)