Create files for the semantic analyzer.
A sudden idea has popped up regarding access specification. Instead of using explicit access specifiers,
why not use PIMPL? This forces the files to be separated, with one implementation and one declaration..
Well, not exactly. Let's use an example

	// MyValue.tul

	sub.ValueHolder holder_;

	(32u cur_val:) getValue
	{
		return holder_.actual_value;
	}

The callee is ensured to not be able to access ValueHolder if that class is put into directories
in the following manner.

	root/
	|--- Enter.tul
	|------ sub/
	|	|--- MyValue.tul
	|	|------ sub/
	|	|	|--- ValueHolder.tul

Now Enter.tul:

	(:) enterProgram
	{
		var sub.MyValue value_;
		32u result_ = value_.getValue()~cur_val;
		sml.String str_ = result_.toString()~string_;
		sys.std_out.printString(str_);
	}

with the ValueHolder.tul being:

	32u actual_value;

	(:) ValueHolder
	{
		actual_value = 100;
	}

There's a small element I dislike, but also like about this. It forces the PIMPL idiom.
However, it becomes somewhat more difficult when attempting to move trees around.
It's probably better to have access specifiers...

In the above examples, we can not do

	var sub.MyValue value_;
	var sub.sub.ValueHolder xval_ = value_.holder_;

Because this breaks the transitivity rule. Neither can value_.holder_ as an expression
be used anywhere in Enter.tul.

I'm not entirely sure about this. Suppose we do this. We remove some boilerplate code
and make the files more terse. This also puts more information in the file hierarchy.
I kind-of like it though. What about moving stuff around?

Suppose I want to move the tree above,... well, since dependencies are structures like
this anyways, I don't see how it will create much of a problem. A problem arises when
the structure has more than a few files, and a lot of the modules depend on
that specific subdirectory. What to do?

	root/
	|--- Enter.tul
	|------ sub/
	|	|--- MyValue.tul
	|	|--- SomeStuff.tul
	|	|--- Monster.tul
	|	|------ sub/
	|	|	|--- ValueHolder.tul

Now what? I wanna move MyValue to a global scope because it's so useful. How should it
be moved? It would be sinful to copy ValueHolder. What if SomeStuff and/or Monster
depend on it? Well I think that's a design issue... You should set up the packages
that you get packages that are actually not dependent on something. It looks great
on the grammar though. Really neat and fits well with the current directory structure
already.