Add the in-expression if statement known as "when".
The reason for not using if inside an expression is to avoid any potential
confusion caused by the following:

	a_ = 1 + 3 * b_ - 500000 / alpha_variable * counter_num % mod_v *
	if ( a_ > 5 )
		callSomthing();
	else
		doNothing();

The above code looks suspiciously like a normal statement, but it's actually
an expression. In some cases, this may obfuscate a really nasty bug.

When we see "when", this is immediately resolved. We know it's part of an
expression.