Examplify nested when expression.
Note that the when expression must have an else clause. If it does not, then the value
is indeterminate. It is not allowed by the grammar to not have an else expression.

The expressions between the ) and else are delimited by those. It is no problem
to have arbitrary expressions inside of these.