Move the entire compiler into an ast directory.
The ast directory (package/namespace) ought to take in characters, present
and endOfCharacters function, and return an abstract syntax tree that has
been analyzed for correctness, with built symbol tables.

It makes little sense to test the tree pruner at the level that it is presented
at. The treepruner can however be tested in a top-level test (essentially an
integration test.).