Add symbol table package.
The symbol tables contain necessary data for and when performing semantic
analysis of the program. It's still not given how to go about, but a
general idea is there.

Starting is simple: generate symbol tables over all modules, for all
variables. Then, check every function invocation to see if the types
match.

That's the first thing to do. I'm not sure if anything other than
type checking is really necessary. Well, there will be purity call
and const correctness checking as well. Let's focus on the types
first though. How to encode these? Considering that types can be
quite arbitrary, it'd be best to generate something easy to work
with. For example,

	var [2, [3, [5, [9, 32u]]]] a_;

It is very useful if we can access this in the table. What's even
more useful if we can access the type returned by an operator such as
the [] operator.

	type(a_[expression]) = '[3, [5, [9, 32u]]]'

It's possible to implement this all as a string...