Add a rudimentary function recognizer in the syntax tree.
Currently in pre-alpha stage. The recognized functions are printed
upon the dtor being called. Later on, recognition of further
elements will need to take place in order to perform semantic
analysis.

Note that the const cast is safe in this case, as the iterator
is not used in a loop or to traverseAdd a rudimentary function recognizer in the syntax tree.