Work on the basic ast.
Recursive functions will have to be defined for the
operations that are to be performed on the ast.

First and foremost, the grammar ought to stabilize.
Without a stable grammar, we have a shaky semantic
analyzer. By that I mean that the grammar must be
complete and no more changes ought to be brought to
the grammar.