Add a new class for handling comments in the stream.
Handling comments was slightly trickly. The question is
how to handle them. For example, it'd be nice with nested
comments (which will be implemented later), but then;
how do we escape comments like "/*" inside strings?

Well, I've currently put the comment mechanism at the very
beginning, such that the parser nor lexer need not see the
comments at all. I initially planned to put it right after
lexing, but it seems like a waste of processing power and
it doesn't become as nice as I would like.

So now, the lexer will also take into account quotation
state. This isn't good from a separation-of-concerns point
of view, but it works and it's very unlikely that other
such constructs (grouping constructs under /* */ and "")
will be introduced.

Next up is to actually make the class work, the states
are there but it's not functional right now.