Add a small example for performing recursive descent table generation.
Tables can be generated on multiple passes through the code.
For example, we can pass once to get a function definitions.
From these definitions, we generate a table.
Now, the nice thing about this is that we can move
the entire tree as a child of that table entry. We thus do not
need to go through all nodes, making this efficient.

Anyhow, I first need to populate a module table. For this I will
need some tools. A module ought to be of the form

	(qname, symbols)

Each of the symbols then have further attributes. We also need
to support some operations to make everything easier.
The symbols table is formed:

	(symbol-name, type-signature, access, state-list)

That's all that's needed for each symbol. With the type and the
access specifier as information to use during semantic analysis.
Each of these symbol names will have a symbol entry.
The state-list contains either statements or nothing. This is
where functions come in.

Should perhaps there be two tables, one for symbols used as
variables and one for functions? That does complicate things a bit

So for each item, we put it into a neat data structure that has
the following functions.

	isPrivate(), isPublic(), isRestricted(), isGlobal(), isGlocal()
	getName(), getSignature()

I think that should cover it... Now the symbol table itself will
need a routine that fetches qualified names (or checks for their
existence). How would this be feasible? Suppose the semantic
analyzer gets this table, it starts at the simplest location which
is the enterProgram function in the module designated by the
compiler. Now we can write some routine to elaborate on this
design:

	doSemanticAnalysis(symbol_table, start_module_name)
		var ptr ModuleTable mod_tab = symbol_table.lookUp(start_module_name)
		// Find the enter function
		// We need to provide parameter types due to overloading
		var ptr MethodTable met_tab = mod_tab.lookUp
			(name_: "enterProgram", returns_: "", takes_: "")

		var #sml.Vector(type_: "MethodTable") met_tables;
		met_tables.pushBack(item_: met_tab);

		do (met_tables.isEmpty() == false)
		{
			var ptr MethodTable method_table = met_tables.popBack();
			met_tables.pushVector
				(vector_: analyzeMethod(method_: method_table);
		}

This is of course much simplified, but it outlines the general algorithm.