Remove the assembly statement in favor of the hack statement.
The hack statement is more versatile, and allows compiler implementors
to do whatever they want. hack acts like a function with a namespace
name. It is a keyword in the grammar. It takes any amount of arguments.
The arguments can be dependent upon the scope (unlike #Generators).

An example of hack is how we can generate assembly:

	hack( assembly_: "MOV r0, r1" );

Note the semicolon. Hack is not an expression. It is a statement.
Hack can also be used to generate code into the lower impl.:

	public (: sml.String to_print) printString
	{
		hack(variable_: "to_print", cppcode_: "std::cout << {0};");
	}

In that case, the compiler is geared to generate C++ code, so what
it does internally is substitute the {0} by to_print. The substitution
type and syntax are also compiler-specific. All that the compiler
receives is an invocation of hack, and its arguments. That's all.
The compiler is free to do whatever it wants. The compiler ought
to specify what is available for hacking.

Hacking allows any platform to be made easily accessible. Combine
this with static if and you got yourself a powerful tool.