Fix RingNumber not including the maximum value.
Add a test to ringnumber to test common edge cases and
general use. RingNumber should be stable now.

It is only defined for the range [0, max]. Numbers below zero
are omitted, although these could be added in the future. I see
no need to complicate this binary counter. A simpler setup would
be to use modulos to normalize the values, but this does not allow
inclusive maximum counting.