Add the alias clause to the syntax.
The alias clause is defined as:

	alias new_name = old_name ;

What it does is that it will recognize the left package name as the right package
name. This is useful in libraries that may have long names. Imagine if you have a
library called simplefastmedialibrary, and you would like to abbreviate that:

	alias sf = simplefastmedialibrary;

And you're done. You can now refer to classes in sf as sf.ClassName. This removes
a tonne of obnoxious code. In addition, it will be easier to work with.