Change the entire var syntax.
The new syntax is slightly different and has come about due to considerations of the
static and variadic elements of a class.

	static { Type name; Type name; ...; Type name; }
	var { Type name; Type name; ...; Type name; }

Both declare data, but static keeps it class-wide. It's a global variable. Statics
clash with functions using static if. For now, static if has been replaced by 'statics if',
which will be changed later. Supposedly, it could be `if [constexpr]` instead, considering
[] is already used as a compile-time delimiter.

For the file-scope, a var/static declares data. `grant` sets a grant list. `alias`
constructs an alias. Anything starting with a ( is a public procedure.