Implement shift and other operators for non-this members.
What this means is that

	sml.Out << Expression;

is not valid syntax. In the case of a class' definition:

	// sml/Out.cxy
	(: Type in) << {
		// code
	}

The immediate class class operators have a higher precedence than the normal expression
operators:

	1 + Aa + Aa + 1

Will parse to ((1 + (Aa + Aa)) + 1)
This may result in some peculiar behaviour:

	3 * Aa + Aa

Will multiply the sum of Aa, I think this needs to be fixed. We can't mess up the
precedence like this.