Fix the line and column counting mechanism.
The original mechanism could report right after the token has ended. Perhaps I should
make it so that it is where the token starts? That seems logical. Anyways. It now
reports from where any token's last character is part of that token.

The thing is, making it count from the start would be quite difficult. In fact, it
would make the token counter do a few things. First, if anything comes along, it needs
to check if it is a symbol, or in a string. The problem with that is that it needs
to hold onto more state. Something that I don't see as desirable.