Implement some basic ideas of table building.
I'm still not certain how to go about this. First, the table needs to be built. My
concern is with how the tree is pruned. It doesn't seem that clean. I wish there was
some more formal mway of pruning the tree. Anyways. After pruning, I iterate the tree.

An idea is to have a table builder for each specific case. Say you get a keyword grant
as the first child, delegate that to a special grant builder. Get a variable declaration,
delegate that...

Currently, the parse tree should ideally be transformed into a perfect abstract syntax
tree. That would be ideal. However, what is the ideal abstract syntax tree? Hopefully
I can prune the parse tree in such a way that it will be perfect. I also need to look
if I can change the grammar slightly - to make it more suited. I won't prematurely
optimize the parsing or lexing. Those are fine (even though the parser uses a binary
tree lookup).

So, we get a 'perfect' AST. Then we need to generate a table from that. Yeah, sounds
like recursive definitions would be best. Just use small classes to identify their
respective sub-trees. How would this look for alias? Alias is one of the simplest
ones. Indeed, it would call upon the type subsystem to do its bidding. Okay... Hmm.

Let's focus on a single thing at a time. We first compute all aliases to create a
table of aliases in a class. What will the table look like? Well, we know an alias
is always a class name, so we get a function:

	getAliasType(string) -> Type

on a table looking like

	MyAlias: sml.ComplicatedClassThatsWayTooLongToNormallyWrite
	AnotherAlias: 3u
	Bool: 1u
	VertexArray: [3, float]

Okay, good. So we now have an alias interface per module. Each module will have all
its aliases in this place. Whenever we then go over any other code, we can look up
the type inside the alias section to check if the type is there.

The problem is that alias can also be used inside a function itself, so a function
locally needs to be able to declare stuff as well. A function will be tied to its
surroundings, but also acts as a single entity. For example, building the tree for
a function can be done independently of the surrounding scope because the function
works on its own. Unresolved types are just delayed and put into a table for unresolved
types. Later on, all the unresolved types are resolved. If they are not, we have an
error. What happens next?

So for functions, we actually need to segment the code into different states. For
instance; we must create a vector of declarations and runnable code in that order:

	[Declaration(s), code, declarations, code, ...]

This is because declaring before use is important. If the code uses something that
is not declared, then we have an error. A function would be able to utilize global
and glocal variables, and these need to come from a bigger context. This should not
be too difficult, as the system can rercursively compile. Though, we can also say
that those are resolved at link time. Anyhow, let's focus on the single class. The
aliases don't care about declaration order, so they will always work. Inside the class,
we create a vector of code and declarations. How about branches? They are their own
little functions - in a sense. They can use what has previously been declared, but
anything declared inside them will not be available outside of the branch. I need
a way of structuring all this information.

	Module:
		Alias Table (AliasName, Type)
		Grant Table (GrantName, Assertions)
		Variable Members (Globality, Visibility, Name, Type)
		Static Members (Globality, Visibility, Name, Type)
		Method Members (Globality, Visibility, Name, Type)
		Visible Globals (Name, Type)

That is how a module will look like. Now let's think of methods of breaking this up.
The alias table is really just a table of names and types. The type will need to be
represented in some kind of vector form:

	Type: [ref, ptr, 32u].

Since types are linear in composition, we can easily perform operations on them. For
example, saying:

	Type.dereference()

Should return a list that has removed the ptr from the beginning. Else, it will throw
an error, because the first type is not a 'ptr'. Enumerations can be used for this.
Perhaps even CrossTerminal, considering their availability. Alright. So types are
pretty easy to work with. Type is just a class. That's neat. What about the other
stuff that's going on?

Grant table with assertions? Hmm, well a GrantName is just a classname, so a string
will suffice. The assertions are a little different. Let's analyze how they would
work. An assertion is basically a function signature or a variable signature. In fact,
these are actually variable/static/method members without their actual content (such
as default assignment, or method definition).

The type of a function is a little different.

	Type: [function]
	Return: [[name, Type], [name2, Type],...]
	In: [[name, Type], [name2, Type],...]

So,... what happens now? What about nested functions? Well we need a "functiontype".
So a type can contain a "functiontype". The functiontype can contain more types, as
it contains a list of in and output parameters. Okay, cool, we can then at least check
if i/o params exist. We also need to overload on the types, such that types are part
of the key to the table.

Let's start with a table of types and take it from there.