Add the most basic expressions to the semantic checker.
The current method of accepting expressions is to check if certain children are any
acceptable expressions. This contains all other expressions, as well as the resources.
This implies a simplification on the high level of the expression tree. The resources
themselves need to check how they operate on the type expressions.