Discuss code generators and comments.
The preprocessor design is complicated due to the code generators as well as the
comments. To remedy this, a design choice was made. To optimize for speed, the code
generation phase comes after the commenting phase. The commenting phase must only
look for `# `, `#(`, and `#)`. It will ignore and copy out verbatim all other instances
of the hash tag. Including `#[`.

The code generator will only recognize openings as `#[`, and closings as `#]`.